Changed scripts to always read in $variables as strings, so that inputs that should be treated as strings, but are actually numbers won't get converted to, say, scientific notation. Also, Static.resolveConstruct should be parsing Longs, not Integers.
This *might* have side effects on scripts that were doing type checking on user input using is_integer, etc, however the impact should be minimal, I would imagine, since there is currently no real use case for doing this.

+++ src/main/java/com/laytonsmith/core/Static.java
@@ -386,7 +386,7 @@ public final class Static {
             return new CBoolean(false, t);
         } else {
             try {
-                return new CInt(Integer.parseInt(val), t);
+                return new CInt(Long.parseLong(val), t);
             } catch (NumberFormatException e) {
                 try {
                     if (val.contains(" ") || val.contains("\t")) {
